pyproject.toml
[project]
name = "labwork4"
version = "0.1.0"
description = "api implementation"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.7",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "pydantic[email]>=2.10.6",
    "pydantic-settings>=2.7.1",
    "sqlmodel>=0.0.22",
    "starlette>=0.45.3",
    "uvicorn>=0.34.0",
    "pyjwt>=2.8.0",
    "asyncpg>=0.29.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "alembic>=1.13.1",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.20",
    "psycopg2[async]>=2.9.10",
]

[dependency-groups]
dev = [
    "pyright>=1.1.392.post0",
    "pytest>=8.3.4",
    "ruff>=0.9.3",
    "pytest-asyncio>=0.23.5",
]
